Refactor history & patches callbacks to updatePatches in both Contacts and Offices
