+++ packages/octave/lib/server/contacts/callbacks/index.js
@@ -2,4 +2,3 @@ export * from './algolia.js'
 export * from './offices.js'
 export * from './projects.js'
 export * from './past-projects.js'
-export * from './patches.js'
