+++ packages/octave/lib/server/offices/callbacks/index.js
@@ -3,4 +3,3 @@ export * from './contacts.js'
 export * from './past-projects.js'
 export * from './projects.js'
 export * from './twilio.js'
-export * from './patches.js'
